Remove IsAvailableAsync check
The function doesn't check if a passcode has been set. So if the user
has set up a passcode (without biometrics) they'll be treated as if they
have no security set up, which is false.

This is a TODO to fix in the future, but for now, it's acceptable to
remove this check since device policies force the user to implement
some kind of passcode.