Fix multiple dimension bugs with soft keyboard
1. OnKeyboardShowing() fires multiple times when minimizing the app. I
found out that it just sends the same CGRect each time, so the fix (this
time) is to only modify the padding if the begin/end frames Y values are
different. Am considering making a MAUI issue on GH.

2. The endFrame.Height value would differ when rotating the device while
the keyboard is open, leading to wonky padding changes on rotation. The
fix (this time) was to use beginFrame's Height instead.