Fix realm code for NoteEntryViewModel
When we implemented this previously we didn't realize that by holding a
valid object you can commit changes automatically and instantly instead
of waiting and saving in one go.

By holding a live Realm object, we don't have to worry about saving
anymore. This also seems to fix the weird performance and bug with
the editor "rewinding" when typing as well. We're no longer holding onto
potentially stale text and overwriting what remained in the Editor.