Move and refactor VisitzRealm into VisitzModel
VisitzKey depends on MAUI for SecureStorage, so it needed to be
untangled from VisitzRealm before it could be moved into VisitzModel.
I took the opportunity to refactor VisitzRealm into a base class impl
instead of copy/pasting everything in one file.

To keep ease-of-use similar to what it was before, I created
VisitzRealms (note the 's') in the Visitz namespace as a wrapper for any
VisitzRealmBase impl.