Force workaround to run on MainThread
Had to use a TaskCompletionSource in case the calling thread wasn't the
main thread, or else the available boolean would not be set in time for
it to be returned.

This solution isn't ideal since it blocks the background thread, but the
Windows network calls seem to work fast anyway.